made changes

+++ main.h
@@ -8,7 +8,7 @@
 /*function names used*/
 int _printf(const char *format, ...);
 int _stringlength(const char *str);
-int convert_to_string(int number);
+int convert_to_string(long int number);
 int convert_to_binary(int value_bin);
 int convert_to_heXa(int value_bin);
 int convert_to_hexa(int value_bin);
